Refactor for better organization
