+++ components/clientComponents/globals/Header/Header.tsx
@@ -13,7 +13,7 @@ import Markdown from "markdown-to-jsx";
 import { useFeatureFlags } from "@lib/hooks/useFeatureFlags";
 import { FeatureFlags } from "@lib/cache/types";
 import { SkipLink } from "../SkipLink";
-import { PublishButton } from "./PublishButton";
+import { PublishButton } from "./PublishMenu";
 type HeaderParams = {
   context?: "admin" | "formBuilder" | "default";
   className?: string;
